Make pfc commands use a class.
This change puts contents originally in pfc/main.py into a class,
to support the usage of the multi-asic helper in a future change.
This change is required, as multi-asic helper relies on certain members
of the class to exist. The multi-asic class helper will be used to add
multi-asic support to pfc commands.