[chassis][multi-asic] Make PFC commands use a class (#3057)
What I did
This change puts contents originally in pfc/main.py into a class, to support the usage of the multi-asic helper in a future change. This change is required, as multi-asic helper being used expects members self.config_db and self.db to exist when a function with the decorator run_on_multi_asic is called. The multi-asic class helper will be used to add multi-asic support to pfc commands in a following pull request.

This is a part of the set of changes being pushed for sonic-net/sonic-buildimage#15148

How I did it
Moved contents of PFC commands into a class. There are no functional changes.

Co-authored-by: rdjeric <rdjeric@arista.com>
Co-authored-by: Kenneth Cheung <kennethcheung@arista.com>