Start lag_keepalive before OA pause, and fork after building packets
Start lag_keepalive before pausing orchagent, so that there's less of a
delay between when orchagent is paused and when kexec happens, and so
that fewer events/changes aren't handled by orchagent.

Additionally, add an option into the lag_keepalive script to fork into
the background after generating the LACPDUs and opening sockets, but
before sending the actual packets. This serves as a sort-of error check
to make sure that it is at least able to send LACPDU packets, and didn't
bail out early.

Signed-off-by: Saikrishna Arcot <sarcot@microsoft.com>