Optimize lag_keepalive by crafting the LACPDU packet ourselves (#3170)
* Optimize lag_keepalive by crafting the LACPDU packet ourselves

Instead of waiting for a LACPDU packet to be sent and capturing that
(which involves waiting roughly 30 seconds), get the necessary
information from teamd and craft it ourselves. This means that the
60-second wait in making sure that a LACPDU packet is captured and the
keepalive script is ready can be largely eliminated (this has been
reduced to 10 seconds to make sure the script has a chance to craft the
packets and send some LACPDUs).

Signed-off-by: Saikrishna Arcot <sarcot@microsoft.com>

* Fix pre-commit errors

Signed-off-by: Saikrishna Arcot <sarcot@microsoft.com>

* Keep a socket open, and reuse that for sending LACPDUs

Signed-off-by: Saikrishna Arcot <sarcot@microsoft.com>

* Add logic to fork into background after collecting information

Signed-off-by: Saikrishna Arcot <sarcot@microsoft.com>

* Start lag_keepalive before OA pause, and fork after building packets

Start lag_keepalive before pausing orchagent, so that there's less of a
delay between when orchagent is paused and when kexec happens, and so
that fewer events/changes aren't handled by orchagent.

Additionally, add an option into the lag_keepalive script to fork into
the background after generating the LACPDUs and opening sockets, but
before sending the actual packets. This serves as a sort-of error check
to make sure that it is at least able to send LACPDU packets, and didn't
bail out early.

Signed-off-by: Saikrishna Arcot <sarcot@microsoft.com>

---------

Signed-off-by: Saikrishna Arcot <sarcot@microsoft.com>