optimization: prevent unneeded deep copies that show up in profiling
Profiling shows copy/deepcopy being a large consumer of resources.

Loading the configuration into sonic-yang-mgmt was performing a deep
copy of the data before loading it because it might be mutated.  With
a recent patch to sonic-yang-mgmt it is no longer mutated so we no
longer need this.

Also when performing a JsonMove apply operation, it does a deepcopy
of the provided data, so this patch moves the JsonMove out of each
independent validator and calls it from the parent that iterates
across validators.