add BulkKeyLevelMoveGenerator()
Add a new Move generator similar to KeyLevelMoveGenerator(), but is
executed before and groups all adds and removes for keys in a table.
It looks for count of pipe (|) delimiters in a key to determine if a new
group should start or not as some tables may have different key
groupings.