Remove invalid test cases
Test cases that attempt to remove the entire path of "" (or /) are invalid.  These
tests are invalid as JsonPatch() does not support such a concept.  While the
validator itself technically does, when it comes to actually applying the patch
with JsonPatch() it fails.  The recent modifications to GCU mean that the validator
needs to take the simulated configuration so we apply the patch before running
the validator which fails.  Testing the original code with no changes other than
applying the patch yields a failure as well, so there is no regression, the tests
are just invalid.