config_db updates: work around race condition
A race condition has been found in the way config_db is updated.
This patch adds a workaround (hack) by simply adding a sleep
for 1s between each patch.  The prior logic had a highly-intensive
sanity check that was removed that took approxmiately the same
amount of time effectively hiding the race condition.