Set the timeout to be larger on VS chassis and speedup a bit (#3861)
What I did
The route_check.py on UpStream LC of virtual chassis takes a longer time to complete while the current timeout limit is 2 mins. Thus, I want to:

Set a larger timeout when the device is a virtual chassis.
Speedup the route_check process a bit by simplifying the string manipulation.