Update sonic-utilities build for Trixie
This includes allowing the use of newer Python modules as well as
updating tests for Python 3.13.

One notable exception is pyroute2, where the minimum version has been
bumped down to 0.7.7. This is so that the Debian package can be reused.
If there are issues with this, then it can be reverted to use 0.7.12 as
long as the Debian package is removed from the slave container.

Signed-off-by: Saikrishna Arcot <sarcot@microsoft.com>