Secureboot: Image signing verification
The DB Key is not guaranteed to be a CARoot.  It is possible to have
the DB Key be an intermediate as it is trusted via the EFI signing
methods which do not use the normal PKI methodology.

This adds a secondary signing verification, the original plus the
above described method.

Signed-off-by: Brad House <bhouse@nexthop.ai>