Skip PFC history test cases on 32-bit architectures due to 32-bit time limits
The version of Python in Debian Bookworm armhf uses 32-bit time, meaning
that timestampts in Python are subject to the Y2038 issue. In Trixie,
Python was compiled with 64-bit time support on armhf, meaning that that
build of Python is no longer subject to the Y2038 issue.

With sonic-utilities tests being reenabled on Bookworm in buildimage,
XFAIL these test cases on 32-bit architectures.

Signed-off-by: Saikrishna Arcot <sarcot@microsoft.com>